Add files via upload
Added an alternative ChessBoard class with the three fields (2 1-D array for the players, 1 2D array that outputs the character spacing on the board) and a main class ChessGame that can properly outputs the board. I figured with this (or even creating a duplicate board to reference the moves, then have the main board output the toString() methods that symbolize the character (if color==black) return (lower case letter); else return (upper case letter), that way we could essentially make the move and display the character on the board? I'd love to get your thoughts! :)